task: Code cleanup
