+++ src/assets/templates/settings.html
@@ -198,6 +198,7 @@
         document.getElementById('sftp2Pass').value = sftp2Config.pass;
       }
     }
+    
     loadSettings();
 
     document.getElementById('smtpForm').addEventListener('submit', async (e) => {
